java

+++ area.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class area {
     static public void main(String args[]){
        Scanner sc=new Scanner(System.in);
-       double r = sc.nextFloat();
+       double r = sc.nextDouble();
        double Area =(3.14*r*r);
        System.out.print(Area);
